+++ Classes/ViewHelpers/Widget/Controller/LanguageMenuController.php
@@ -33,9 +33,11 @@ namespace KayStrobach\Themes\ViewHelpers\Widget\Controller;
  * @author Thomas Deuling <typo3@coding.ms>
  * @package themes
  */
-use SJBR\StaticInfoTables\Domain\Repository\LanguageRepository;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Query\Restriction\FrontendRestrictionContainer;
+use TYPO3\CMS\Core\Site\Entity\Site;
+use TYPO3\CMS\Core\Site\Entity\SiteLanguage;
+use TYPO3\CMS\Core\Site\SiteFinder;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
@@ -49,116 +51,59 @@ class LanguageMenuController extends \TYPO3\CMS\Fluid\Core\Widget\AbstractWidget
     protected $configuration = [];
 
     /**
-     * Language Repository.
-     *
-     * @var \SJBR\StaticInfoTables\Domain\Repository\LanguageRepository
-     * @inject
+     * @var \TYPO3\CMS\Core\Site\SiteFinder
      */
-    protected $languageRepository;
+    protected $siteFinder = null;
 
     /**
-     * ContentObjectRenderer.
-     *
-     * @var \TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer
-     * @inject
-    protected $contentObjectRenderer;
-*/
-
-    /**
-     * @param \SJBR\StaticInfoTables\Domain\Repository\LanguageRepository $languageRepository
-     *
-     * @return void
+     * @param \TYPO3\CMS\Core\Site\SiteFinder $siteFinder
      */
-    public function injectLanguageRepository(LanguageRepository $languageRepository)
+    public function injectSiteFinder(SiteFinder $siteFinder)
     {
-        $this->languageRepository = $languageRepository;
+        $this->siteFinder = $siteFinder;
     }
 
     /**
      * @return void
+     * @throws \TYPO3\CMS\Core\Exception\SiteNotFoundException
      */
     public function indexAction()
     {
         $menu = [];
-        $availableLanguages = explode(',', '0,'.$this->widgetConfiguration['availableLanguages']);
-        $availableLanguages = array_unique($availableLanguages);
-        $currentLanguageUid = (int) $this->widgetConfiguration['currentLanguageUid'];
-        $defaultLanguageIsoCodeShort = $this->widgetConfiguration['defaultLanguageIsoCodeShort'];
-        $defaultLanguageLabel = $this->widgetConfiguration['defaultLanguageLabel'];
-        $defaultLanguageFlag = $this->widgetConfiguration['defaultLanguageFlag'];
-        $flagIconPath = $this->widgetConfiguration['flagIconPath'];
-        $flagIconFileExtension = $this->widgetConfiguration['flagIconFileExtension'];
-
-        if (!empty($availableLanguages)) {
-            foreach ($availableLanguages as $languageUid) {
-                $menuEntry = [];
-                $menuEntry['L'] = $languageUid;
-                $menuEntry['pageUid'] = $GLOBALS['TSFE']->id;
-                $class = 'unknown';
-                $label = 'unknown';
-                $flag = 'unknown';
-                $hasTranslation = true;
-
-                // Is active language
-                $menuEntry['active'] = ((int) $currentLanguageUid === (int) $languageUid);
-
-                // Is default language
-                if ((int) $languageUid === 0) {
-                    $class = $defaultLanguageIsoCodeShort != '' ? $defaultLanguageIsoCodeShort : 'en';
-                    $label = $defaultLanguageLabel != '' ? $defaultLanguageLabel : 'English';
-                    $flag = $defaultLanguageFlag != '' ? $defaultLanguageFlag : 'gb';
-                } elseif (($sysLanguage = $this->getSysLanguage($languageUid))) {
-                    if (!($this->languageRepository instanceof LanguageRepository)) {
-                        $this->languageRepository = GeneralUtility::makeInstance('SJBR\\StaticInfoTables\\Domain\\Repository\\LanguageRepository');
-                    }
-                    $languageObject = $this->languageRepository->findByIdentifier($sysLanguage['static_lang_isocode']);
-                    if ($languageObject instanceof \SJBR\StaticInfoTables\Domain\Model\Language) {
-                        $class = $languageObject->getIsoCodeA2();
-                        $label = $languageObject->getNameEn();
-                    }
-                    $flag = $sysLanguage['flag'];
-                    $hasTranslation = $this->hasTranslation($GLOBALS['TSFE']->id, $languageUid);
+        $pageUid = (int)$GLOBALS['TSFE']->id;
+        $languageUid = (int)$GLOBALS['TSFE']->sys_language_uid;
+        /** @var Site $siteConfiguration */
+        $siteConfiguration = $this->siteFinder->getSiteByPageId($pageUid);
+        /** @var SiteLanguage $language */
+        foreach($siteConfiguration->getAllLanguages() as $language) {
+            if($language->isEnabled()) {
+                //
+                $item = [
+                    'L' => $language->getLanguageId(),
+                    'pageUid' => $pageUid,
+                    'active' => $languageUid === $language->getLanguageId(),
+                ];
+                //
+                // Navigation title, with fallback on title
+                $item['label'] = $language->getNavigationTitle();
+                //
+                // CSS class and flag
+                $item['class'] = strtolower($language->getTwoLetterIsoCode());
+                $item['flag'] = strtoupper($language->getTwoLetterIsoCode());
+                $item['flagIdentifier'] = $language->getFlagIdentifier();
+                //
+                // Has translation?
+                $item['hasTranslation'] = true;
+                if($language->getLanguageId() > 0) {
+                    $item['hasTranslation'] = $this->hasTranslation($pageUid, $language->getLanguageId());
                 }
-
-                $menuEntry['label'] = $label;
-                $menuEntry['class'] = strtolower($class);
-                $menuEntry['flag'] = strtoupper($flag);
-                $menuEntry['hasTranslation'] = $hasTranslation;
-                $menu[] = $menuEntry;
+                //
+                $menu[] = $item;
             }
         }
-
         $this->view->assign('menu', $menu);
-        $this->view->assign('flagIconPath', $flagIconPath);
-        $this->view->assign('flagIconFileExtension', $flagIconFileExtension);
-    }
-
-    /**
-     * Get the data of a sys language.
-     *
-     * @param $uid \int Language uid
-     *
-     * @return \array Language data array
-     */
-    protected function getSysLanguage($uid = 0)
-    {
-        $sysLanguage = array();
-        /** @var \TYPO3\CMS\Core\Database\Query\QueryBuilder $queryBuilder */
-        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
-            ->getQueryBuilderForTable('sys_language');
-        $queryBuilder->select('uid', 'title', 'flag', 'static_lang_isocode')
-            ->from('sys_language')
-            ->where(
-                $queryBuilder->expr()->eq(
-                    'uid', $queryBuilder->createNamedParameter((int) $uid, \PDO::PARAM_INT)
-                )
-            );
-        /** @var  \Doctrine\DBAL\Driver\Statement $statement */
-        $statement = $queryBuilder->execute();
-        if ($statement->rowCount()>0) {
-            $sysLanguage = $statement->fetch();
-        }
-        return $sysLanguage;
+        $this->view->assign('flagIconPath', $this->widgetConfiguration['flagIconPath']);
+        $this->view->assign('flagIconFileExtension', $this->widgetConfiguration['flagIconFileExtension']);
     }
 
     /**
@@ -169,18 +114,21 @@ class LanguageMenuController extends \TYPO3\CMS\Fluid\Core\Widget\AbstractWidget
      *
      * @return bool
      */
-    protected function hasTranslation($pid, $languageUid)
+    protected function hasTranslation($pid, $languageUid): bool
     {
         $hasTranslation = false;
+        /** @var ConnectionPool $connectionPool */
+        $connectionPool = GeneralUtility::makeInstance(ConnectionPool::class);
         /** @var \TYPO3\CMS\Core\Database\Query\QueryBuilder $queryBuilder */
-        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
-            ->getQueryBuilderForTable('pages');
-        $queryBuilder->setRestrictions(GeneralUtility::makeInstance(FrontendRestrictionContainer::class));
+        $queryBuilder = $connectionPool->getQueryBuilderForTable('pages');
+        /** @var FrontendRestrictionContainer $frontendRestrictionContainer */
+        $frontendRestrictionContainer = GeneralUtility::makeInstance(FrontendRestrictionContainer::class);
+        $queryBuilder->setRestrictions($frontendRestrictionContainer);
         $queryBuilder->select('uid')
             ->from('pages')
             ->andWhere(
                 $queryBuilder->expr()->eq(
-                    'pid', $queryBuilder->createNamedParameter((int) $pid, \PDO::PARAM_INT)
+                    'l10n_parent', $queryBuilder->createNamedParameter((int) $pid, \PDO::PARAM_INT)
                 ),
                 $queryBuilder->expr()->eq(
                     'sys_language_uid', $queryBuilder->createNamedParameter((int) $languageUid, \PDO::PARAM_INT)